mongodb: fix warning about loss of precision
git-svn-id: https://svn.osgeo.org/gdal/trunk@36662 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/mongodb/ogrmongodbdriver.cpp
@@ -1091,7 +1091,7 @@ static void OGRMongoDBReaderSetField( OGRLayer* poLayer,
                     double dfVal = elt.Double();
                     if( dfVal < std::numeric_limits<GIntBig>::min() )
                         panValues[i] = std::numeric_limits<GIntBig>::min();
-                    else if( dfVal > std::numeric_limits<GIntBig>::max() )
+                    else if( dfVal > static_cast<double>(std::numeric_limits<GIntBig>::max()) )
                         panValues[i] = std::numeric_limits<GIntBig>::max();
                     else
                         panValues[i] = (int)dfVal;
